+++ pidman/urls.py
@@ -9,6 +9,7 @@ urlpatterns = [
     # pidman has no user-facing index, so redirect from base site url
     # to site admin page
     url(r'^$', RedirectView.as_view(url='admin/', permanent=False)),
+    url(r'^admin/linkcheck/', include('linkcheck.urls', namespace='linkcheck')),
     url(r'^admin/', include(admin_site.urls)),
 
     # REST API urls live under top-level, but should not conflict with resolver urls